bot-engine/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
ccxt==4.3.56
pandas==2.2.2
ta==0.11.0
apscheduler==3.10.4
asyncpg==0.29.0
sqlalchemy==2.0.31
python-dotenv==1.0.1
cryptography==42.0.8
pycryptodome==3.19.0
httpx==0.27.0
websockets==12.0
pydantic==2.7.4
pytz==2024.1